Class

+++ Class/Class.Node/Create.cs
@@ -3139,8 +3139,8 @@ public class Create : InfraCreate
 
     protected virtual Node ExecuteList(NodeKind kind, RangeState rangeState, NodeState nodeState, Range range)
     {
-        int start;
-        int end;
+        long start;
+        long end;
         start = range.Start;
         end = range.End;
